raw_python/requirements.txt
pillow
piexif
python-dotenv